add DetailItemViewModel
